requirements.txt
streamlit
openai-whisper
requests
python-dotenv
google-generativeai